Added support for parsing out template frontmatter
Frontmatter can be specified in a json-like manner. It is relatively lenient, but commas must be specified to separate key->value pairs. The outermost wrapping curly braces have been replaced with wrapping "---" - a sequence of three dashes - as is done for Jekyll templates.

Keys need not be wrapped in quotation marks, but string values MUST be wrapped as such.

Arrays and objects work just as in normal json, except for the outermost braces.

For controllers, Standard keys and their usages are:

title: Title of the action name
description: Description of the action name
finders: Finder methods that need to be set for the model
related: Related finder methods that need to be set for the model

Using frontmatter is a neat way of specifying extra data that can be used by other templates further on down the compilation step. It can be thought of as a really bad implementation of evented programming.